requirements.txt
tensorflow==2.6.0
opencv-python
pillow